+++ include/power.h
@@ -3,13 +3,10 @@
 
 #include <Arduino.h>
 
-// Disconnects MQTT/WiFi and puts the chip into deep sleep for the given
-// duration. Never returns - the chip resets on wake and setup() runs
-// again from the top.
+/* Disconnects MQTT/WiFi and puts the chip into deep sleep for the given duration. Never returns - the chip resets on wake and setup() runs again from the top. */
 void goToSleep(uint64_t microseconds);
 
-// Blinks the LED, then restores it to ON (only ever called while we
-// already know the switch is on).
-void blinkLED(uint8_t times, uint16_t intervalMs);
+/* Blinks the LED, then restores it to ON (only ever called while we already know the switch is on). */
+void blinkLED(int times, uint16_t intervalMs);
 
 #endif // POWER_H
